requirements.txt
django==5.0
djangorestframework~=3.14.0
djangorestframework-simplejwt==5.3.1

requests~=2.31.0
responses~=0.24.1
markdown~=3.5.1
django-filter==23.5
django-environ

celery[redis]~=5.3.6

psycopg2-binary

drf-spectacular
pre-commit
flake8~=6.1.0
black~=23.11.0

django-debug-toolbar
django-extensions
sentry-sdk[django]

pytest~=7.4.3
coverage
django-phonenumber-field[phonenumberslite]
